Fix an uncomfortable situation.Read description
When we firstly place the agvs/loads/end point the path can be placed
OVER other agv's start point.theoritically it cant happen.So...it now
creates the path considering the agv's start points as
non-walkables.when the START button is pressed,this is reverted ,finding
the correct path and works as it is supposed to.

-Why we did this?
we may need ,one day,do not start an agv from the beginning.