functions.cs: minor bug fix (details @ descr.)
Scenario with many AGVs: If a vehicle was trapped in an area with Loads
in its range, then those loads were marked as available. Wrong, because
other AGVs would finish their tasks and then would try to collect the
trapped Loads leading to an endless loop of AGVs creating path to the
top left of the grid and then back to the end.