project:Hack into TreeView to add DoubleBuffered for it
The issue:
Updating TreeView was causing a flickering effect.The reason is that TreeView does NOT support Double buffered graphics.One solution was ,to simply override the .Clear()/Refresh() functions and make them do nothing by creating a Message Callback with get;set; which would return null.When i tryed this , the UI was lagging VERY MUCH because we are using Refresh/invalidate many times inside the code.

The solution:
Since i cant override the functions refresh/invalidate ,i override the WHOLE TreeView Element.
3 const values are describing the ASM code for dll calling wrapper on TreeViews. (more infos here https://msdn.microsoft.com/en-us/library/windows/desktop/ms644950(v=vs.85).aspx )

Dll import is needed for SendMessage API call.

OnHandleCreated: same logic as form_load.When the element is created,send the custom message.

Before that event,the CreateParams override ensures that tree's style is "hackable"

The WndProc override is a failsafe solution just in case the above code wont work.If the above code fail, try to grab the callbacked message and return NULL as respond.that will cause internal crash and wont do anything