functions:initialization:Extend BaseGrid as StaticGrid instead of DynamicGridWPool
We have 4 types of grids:
StaticGrid->The default Grid.as the name says its static.It marks the walkables/unwalkables as static so they dont update after scans.We dont care about that since we manually scan.
DynamicGrid->The opposite situation compared to staticgrid.
DynamicGridWpool->Uses the same node pool for all processes.That was our default since the beginning.
PartialGridWPool->Partial grid allow to make scans in a specific area.We surely cant use that

DynamicGridWPool vs StaticGrid (our situation):
-------------------------------------------------------
DynamicGridWPool is supposed to be more accurate but costs a lot on CPU usage (almost 90% when load is placed).The con here is ,that DynamicGridWPool can dynamically handle the walkable blocks.Although we are already doing this with our custom algorithms.So the high cpu usage for that process is useless for us.
In conclusion, simply switch to staticgrid to decrease the CPU usage (almost 70% on load place).The bottleneck here is,that staticgrid uses more memory ( ~100mb more)