functions.cs: [BUG fix] inside FullyRestore -> desc
SearchGrid was not reinitialized correctly. When the grid was getting
resized, old SearchGrid would still hold its old values for cells
(walkable or not). This had as a result the A* searching in a
differently sized SearchGrid than what it really was, causing troubles
in arrays etc etc (overflows). Damn chain reaction....