functions.cs: Optimized CheckForTrappedLoads
Now receives a second parameter, the endPoint as GridPos and uses it to
check if a load can be driven to the exit. IF so, then the whole check
process stops because we don't care to know what's the state of ALL the
next loads.
This way we save time and performance.