main_form:Discover the MouseWheel event
the "problem":
We wanted to replace the ctrl+"press a button" to resize the grid ,with a 'ctrl+scroll up/down' solution.Although we thought that the main form does not "support" a mouse wheel event.

the "workaround":
first of all, we couldnt 'believe' that there is no mousewheel event.Searching in the C# source code we found that the event is there but visual studio doesnt show it at the properties.
reference:https://referencesource.microsoft.com/#System.Windows.Forms/winforms/Managed/System/WinForms/Control.cs,5129

Browsable=false.

But why ?

The reason is that ,this event requires the Control to be focused //this.Focus()// in contrast to the other events like mouseup etc...

the "solution":
Dynamically declare the event for future use , create the callback function and focus the control.